create-config suitable for publishing via public github pages
There are several moving parts to this configuration machine.

I want to publish a configuration on any web site for simplicity.
Most of the configuration is public info anyway. But there are three
secrets that I want to encrypt.

I'll generate an RSA key pair on the server to use for encrypting the
secrets. That's what remote-keygen does. Well, it also saves a copy of
the public key in our ssh config folder so we can use it to encrypt
the secrets.

create-config builds a structured configuration, encrypting the
secrets. Soon we will build the deploy code which consumes this config
file and decrypts the secrets on the server.