create separate configuration for production and staging domains
Motivated by an experience with Let's Encrypt rate limits.

deploy/compile-config includes extra complexity in the jq expression
to handle a case where we don't yet have staging domains configured.
Specifically we could simplify this expression if we knew there would
always be staging domains in the configuration:
    (.staging//[])[]
to just this
    .staging[]