add instructions to configure ssh access for a new workstation
This is fiddly and involves switch contexts between the workstation,
the digital ocean console, and ssh-ing to multiple accounts. It might
warrant writing some scripts to reduce the fiddling, but for my own
use case, I anticipate only needing to do this once every five years
or so. A lot could change on the internet between now and then, so we
choose to leave it at a documentation update.